Actually emit multiple errors when parsing one statement fails
The code was missing the necessary plumbing to emit mutliple errors.
That was silly because I have put hours in the code to make sure that we
do not halt parsing at the first sight of error. This patch allows us to
parse even when there is an error in the source.

When the parser sees the next error, it skips that line and jumps to the
next one hoping that we'll get to see a sane statement there.