Don't throw exceptions while parsing
Even fatal exception should set a flag that  tells the parser that we
cannot parse any further. Whenever a call to `parseNextStatement`
occurs, we should check if we failed to parse the next statement and if
we did, can we parse further ahead. If we cannot, then we should give up
parsing. This patch adds that functionality.